feat: ExtraFields.UseJSONTag
if ExtraFields.UseJSONTag = true, the json annotation (if defined)
on struct is used to name the key.

The KeyFormatterFunc can now use the level of the key
a.b
a is on level 0, b on level 1.

It's now possible to write a formatter like that:

```go
e.Formatters = []dump.KeyFormatterFunc{
	func(s string, level int) string {
		if level == 0 {
			return strings.ToLower(s)
		}
		return s
	},
}
```

Signed-off-by: Yvonnick Esnault <yvonnick.esnault@corp.ovh.com>